Test runner refactor
This commit breaks the fixure loading into it's own utility and
does a better job of managing application context during tests.

The `Loader` class accepts an application object, a database, and
an array of fixtures. `Loader.load()` will iterate through the fixtures
and add structured data into an in-memory sqlite database for the
tests.

Other notes:
- `/auto_login` has been moved into the `TestBase` object for brevity.
- The user object is returned to the test following login
- The only ORM supported at this time is SQLAlchemy.